+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Change: Regexes/Patterns allow more characters like `/` and `!` to be
 used directly now. But only in the context of a single regex
 pattern like `$r(foo/bar)`. This makes it easier to process text.
+* Feature: Global regex matches with `$rg/.../`.
+* Feature: Global regex substitutions with `$rs/.../`.
 
 0.6.3 (2020-08-06)
 ==================